requirements.txt

cuda90=1.0
cudatoolkit=8.0
cudnn=7.0.5

matplotlib=3.0.0
numpy==1.15.2
pandas=0.20.3
pillow=5.2.0
opencv-python==4.1.1.26
python=3.5.4
pytorch=1.0.0
scikit-image=0.13.0
scipy=1.1.0
tensorflow-gpu=1.5.0
torchvision=0.2.1
tqdm=4.15.0
h5py==2.8.0


